LayoutTests/fast/css-grid-layout/grid-columns-rows-get-set.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<link href="resources/grid.css" rel="stylesheet">
<style>
.grid {
    /* Give an explicit size to the grid so that percentage grid tracks have a consistent resolution */
    width: 800px;
    height: 600px;
}
.gridItem {
    -webkit-grid-column: 1;
    -webkit-grid-row: 1;
    width: 7px;
    height: 11px;
}
.gridItem2 {
    -webkit-grid-column: 1;
    -webkit-grid-row: 1;
    width: 17px;
    height: 3px;
}
.gridWithNone {
    -webkit-grid-template-columns: none;
    -webkit-grid-template-rows: none;
}
.gridWithFixed {
    -webkit-grid-template-columns: 10px;
    -webkit-grid-template-rows: 15px;
}
.gridWithPercent {
    -webkit-grid-template-columns: 53%;
    -webkit-grid-template-rows: 27%;
}
.gridWithAuto {
    -webkit-grid-template-columns: auto;
    -webkit-grid-template-rows: auto;
}
.gridWithEM {
    -webkit-grid-template-columns: 10em;
    -webkit-grid-template-rows: 15em;
    font: 10px Ahem;
}
.gridWithViewPortPercentage {
    -webkit-grid-template-columns: 8vw;
    -webkit-grid-template-rows: 10vh;
}
.gridWithFitContent {
    -webkit-grid-template-columns: -webkit-fit-content;
    -webkit-grid-template-rows: -webkit-fit-content;
}
.gridWithFitAvailable {
    -webkit-grid-template-columns: -webkit-fit-available;
    -webkit-grid-template-rows: -webkit-fit-available;
}
.gridWithMinMax {
    -webkit-grid-template-columns: minmax(10%, 15px);
    -webkit-grid-template-rows: minmax(20px, 50%);
}
.gridWithMinContent {
    -webkit-grid-template-columns: -webkit-min-content;
    -webkit-grid-template-rows: -webkit-min-content;
}
.gridWithMaxContent {
    -webkit-grid-template-columns: -webkit-max-content;
    -webkit-grid-template-rows: -webkit-max-content;
}
.gridWithFraction {
    -webkit-grid-template-columns: 1fr;
    -webkit-grid-template-rows: 2fr;
}
.gridWithCalc {
    -webkit-grid-template-columns: calc(150px);
    -webkit-grid-template-rows: calc(75px);
}
.gridWithCalcComplex {
    -webkit-grid-template-columns: calc(50% + 150px);
    -webkit-grid-template-rows: calc(65% + 75px);
}
.gridWithCalcInsideMinMax {
    -webkit-grid-template-columns: minmax(10%, calc(15px));
    -webkit-grid-template-rows: minmax(calc(20px), 50%);
}
.gridWithCalcComplexInsideMinMax {
    -webkit-grid-template-columns: minmax(10%, calc(50% + 15px));
    -webkit-grid-template-rows: minmax(calc(20px + 10%), 50%);
}
</style>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div class="grid gridWithNone" id="gridWithNoneElement"></div>
<div class="grid gridWithFixed" id="gridWithFixedElement"></div>
<div class="grid gridWithPercent" id="gridWithPercentElement"></div>
<div class="indefiniteSizeGrid gridWithPercent" id="gridWithPercentWithoutSize"></div>
<div class="indefiniteSizeGrid gridWithPercent" id="gridWithPercentWithoutSizeWithChildren">
    <div class="gridItem"></div>
</div>
<div class="grid gridWithAuto" id="gridWithAutoElement"></div>
<div class="grid gridWithAuto" id="gridWithAutoWithChildrenElement">
  <div class="gridItem"></div>
</div>
<div class="grid gridWithEM" id="gridWithEMElement"></div>
<div class="grid gridWithViewPortPercentage" id="gridWithViewPortPercentageElement"></div>
<div class="grid gridWithFitContent" id="gridWithFitContentElement"></div>
<div class="grid gridWithFitAvailable" id="gridWithFitAvailableElement"></div>
<div class="grid gridWithMinMax" id="gridWithMinMaxElement"></div>
<div class="grid gridWithMinContent" id="gridWithMinContentElement"></div>
<div class="grid gridWithMinContent" id="gridWithMinContentWithChildrenElement">
    <div class="gridItem"></div>
    <div class="gridItem2"></div>
</div>
<div class="grid gridWithMaxContent" id="gridWithMaxContentElement"></div>
<div class="grid gridWithMaxContent" id="gridWithMaxContentWithChildrenElement">
    <div class="gridItem"></div>
    <div class="gridItem2"></div>
</div>
<div class="grid gridWithFraction" id="gridWithFractionElement"></div>
<div class="grid gridWithCalc" id="gridWithCalcElement"></div>
<div class="grid gridWithCalcComplex" id="gridWithCalcComplexElement"></div>
<div class="grid gridWithCalcInsideMinMax" id="gridWithCalcInsideMinMaxElement"></div>
<div class="grid gridWithCalcComplexInsideMinMax" id="gridWithCalcComplexInsideMinMaxElement"></div>
<script src="resources/grid-definitions-parsing-utils.js"></script>
<script src="resources/grid-columns-rows-get-set.js"></script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
